refactor around html module
