+++ crates/nu-command/Cargo.toml
@@ -81,7 +81,6 @@ uuid = { version = "0.8.2", features = ["v4"] }
 which = { version = "4.2.2", optional = true }
 reedline =  { version = "0.4.0", features = ["bashisms"]}
 wax = { version =  "0.4.0", features = ["diagnostics"] }
-zip = { version="0.5.9", optional = true }
 
 [target.'cfg(unix)'.dependencies]
 umask = "1.0.0"
