+++ crates/nu-command/src/filters/zip.rs
